* commit '32f44331ead051a71005e957a7da5cdd200a5df2':
  Updated cleanup variable name to debug to be more generally applicable when debugging executors
  Removed unnecessary temporary list of datasets
  WOLF-91 added attribute to allow developers to prevent cleanup of working directories.

+++ datawolf-core/src/main/java/edu/illinois/ncsa/datawolf/LocalExecutor.java
@@ -151,11 +151,13 @@ public abstract class LocalExecutor extends Executor implements Runnable {
             setState(State.FAILED);
             println("Error during execution.", thr);
         } finally {
-            // do some cleanup
-            if ((cwd != null) && !deleteDirectory(cwd)) {
-                String msg = String.format("Could not remove directory [%s].", cwd.getAbsolutePath());
-                logger.info(msg);
-                println(msg);
+            // do some cleanup unless in debug mode
+            if (!isDebug()) {
+                if ((cwd != null) && !deleteDirectory(cwd)) {
+                    String msg = String.format("Could not remove directory [%s].", cwd.getAbsolutePath());
+                    logger.info(msg);
+                    println(msg);
+                }
             }
             flushLog();
         }
